@@ -1,8 +1,8 @@
 package org.notima.generic.businessobjects;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
 import org.notima.generic.ifacebusinessobjects.FactoringReservation;
 
 /**
