@@ -2,8 +2,8 @@ package org.notima.generic.businessobjects;
 
 import java.util.List;
 
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement(name = "BusinessPartnerList")
 public class BusinessPartnerList<B> {
