@@ -2,8 +2,8 @@ package org.notima.generic.businessobjects;
 
 import java.util.Date;
 
-import javax.persistence.Entity;
-import javax.persistence.ManyToOne;
+import jakarta.persistence.Entity;
+import jakarta.persistence.ManyToOne;
 
 import org.notima.generic.ifacebusinessobjects.OrderInvoiceLine;
 
