@@ -3,7 +3,7 @@ package org.notima.generic.businessobjects;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlElement;
 
 public class PaymentWriteOffs {
 
