@@ -1,7 +1,7 @@
 package org.notima.generic.businessobjects;
 
-import javax.persistence.Entity;
-import javax.persistence.ManyToOne;
+import jakarta.persistence.Entity;
+import jakarta.persistence.ManyToOne;
 
 @Entity
 public class PriceListLine {
