@@ -2,7 +2,7 @@ package org.notima.generic.businessobjects;
 
 import java.util.List;
 
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement(name = "ProductInfoBatch")
 public class ProductInfoBatch {
