@@ -2,7 +2,7 @@ package org.notima.generic.businessobjects;
 
 import java.util.Date;
 
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 /**
  * Time record
