@@ -1,6 +1,6 @@
-@javax.xml.bind.annotation.XmlSchema ( 
+@jakarta.xml.bind.annotation.XmlSchema ( 
 		namespace = "https://github.com/notima/businessobjects",
-		elementFormDefault=javax.xml.bind.annotation.XmlNsForm.QUALIFIED
+		elementFormDefault=jakarta.xml.bind.annotation.XmlNsForm.QUALIFIED
 )
 
 package org.notima.generic.businessobjects;
