@@ -2,7 +2,7 @@ package org.notima.generic.businessobjects.test;
 
 import java.time.LocalDate;
 
-import javax.xml.bind.JAXB;
+import jakarta.xml.bind.JAXB;
 
 import org.junit.Test;
 import org.notima.generic.businessobjects.AccountStatementLines;
@@ -11,7 +11,7 @@ public class TestCreateAccountStatementLines {
 
 	@Test
 	public void testCreateAccountStatementLines() {
-
+		
 		LocalDate from = java.time.LocalDate.of(2018, 1, 1);
 		LocalDate until = LocalDate.of(2018, 12, 31);
 		
