@@ -2,7 +2,7 @@ package org.notima.generic.businessobjects.test;
 
 import java.io.PrintWriter;
 
-import javax.xml.bind.JAXB;
+import jakarta.xml.bind.JAXB;
 
 import org.junit.After;
 import org.junit.Before;
